Добавлен микросервис FileStorage. использующий MinIO, и клиентский API для него. Реализовано добавление и хранение фото при добавлении нового обяьвления и изменении старого.

+++ src/Board/Application/Board.Application.AppData/Contexts/Adverts/Services/AdvertService.cs
@@ -1,8 +1,13 @@
 ﻿using AutoMapper;
+using Board.Application.AppData.Contexts.AdvertImages.Repositories;
 using Board.Application.AppData.Contexts.Adverts.Repositories;
 using Board.Application.AppData.Contexts.Adverts.Services;
+using Board.Application.AppData.Contexts.Files.Services;
+using Board.Application.AppData.Contexts.Users.Services;
+using Board.Contracts.Contexts.AdvertImages;
 using Board.Contracts.Contexts.Adverts;
 using Board.Domain;
+using FluentValidation;
 using Microsoft.AspNetCore.JsonPatch;
 using Microsoft.EntityFrameworkCore.Infrastructure.Internal;
 using System;
@@ -16,12 +21,23 @@ namespace Board.Application.AppData.Contexts.Adverts.Services
     public class AdvertService : IAdvertService
     {
         private readonly IAdvertRepository _advertRepository;
+        private readonly IAdvertImageRepository _advertImageRepository;
+        private readonly IUserService _userService;
+        private readonly IFileService _fileService;
         private readonly IMapper _mapper;
+        private readonly IValidator<AdvertAddRequest> _advertAddValidator;
+        private readonly IValidator<AdvertUpdateRequest> _advertUpdateValidator;
 
-        public AdvertService(IAdvertRepository advertRepository, IMapper mapper)
+        public AdvertService(IAdvertRepository advertRepository, IMapper mapper, IValidator<AdvertAddRequest> advertAddValidator,
+            IValidator<AdvertUpdateRequest> advertUpdateValidator, IAdvertImageRepository advertImageRepository, IUserService userService, IFileService fileService)
         {
             _advertRepository = advertRepository;
             _mapper = mapper;
+            _advertAddValidator = advertAddValidator;
+            _advertUpdateValidator = advertUpdateValidator;
+            _advertImageRepository = advertImageRepository;
+            _userService = userService;
+            _fileService = fileService;
         }
 
         public Task<IReadOnlyCollection<AdvertSummary>> GetAllAsync(int? offset, int? count, CancellationToken cancellation)
@@ -44,25 +60,53 @@ namespace Board.Application.AppData.Contexts.Adverts.Services
             return _advertRepository.GetAllFilteredAsync(request, offset.GetValueOrDefault(), count.GetValueOrDefault(), cancellation);
         }
 
-        public Task<AdvertDetails> GetByIdAsync(Guid id, CancellationToken cancellation)
+        public async Task<AdvertDetails> GetByIdAsync(Guid id, CancellationToken cancellation)
         {
-            return _advertRepository.GetByIdAsync(id, cancellation);
+            var advert = await _advertRepository.GetByIdAsync(id, cancellation);
+            var user = await _userService.GetById(advert.UserId, cancellation);
+            advert.User = user;
+
+            return advert;
         }
 
         public async Task<Guid> CreateAsync(AdvertAddRequest addRequest, CancellationToken cancellation)
         {
             // TODO: валидация
+            var validationResult = _advertAddValidator.Validate(addRequest);
+            if (!validationResult.IsValid)
+            {
+                throw new ArgumentException(validationResult.ToString("~"));
+            }
 
             // TODO: логирование
 
-            // TODO: добавить все картинки обьявления, а если потом обьявление не добавится, то удалить их
 
-            //addRequest.UserId = CurrentUser()
+            var currentUser = await _userService.GetCurrent(cancellation);
+            if(currentUser == null)
+            {
+                throw new KeyNotFoundException();
+            }
+            addRequest.UserId = currentUser.Id.GetValueOrDefault();
+
+            // проверка наличия фото с указанными ID на сервере
+            foreach (var imageId in addRequest.AdvertImagesId)
+            {
+                var fileInfo = _fileService.GetInfoAsync(imageId, cancellation);
+                if(fileInfo == null)
+                {
+                    throw new KeyNotFoundException();
+                }
+            }
+
             var newAdvertId = await _advertRepository.AddAsync(addRequest, cancellation);
 
+            foreach(var imageId in addRequest.AdvertImagesId)
+            {
+                var imageAddRequest = new AdvertImageAddRequest { AdvertId = newAdvertId, FileId = imageId };
+                await _advertImageRepository.AddAsync(imageAddRequest, cancellation);
+            }
+            
             // TODO: если обьявление не добавилось, то удалить картинки
-
-            // TODO: обновить информацию о пользователе
 
             return newAdvertId;
         }
@@ -75,14 +119,38 @@ namespace Board.Application.AppData.Contexts.Adverts.Services
         public async Task<AdvertDetails> UpdateAsync(Guid id, AdvertUpdateRequest updateRequest, CancellationToken cancellation)
         {
             // TODO: валидация
+            var validationResult = _advertUpdateValidator.Validate(updateRequest);
+            if (!validationResult.IsValid)
+            {
+                throw new ArgumentException(validationResult.ToString("~"));
+            }
 
             // TODO: логирование
+            var currentAdvert = await _advertRepository.GetByIdAsync(id, cancellation);
+            var currentUser = await _userService.GetCurrent(cancellation);
+            if (currentUser.Id != currentAdvert.UserId && currentUser.RoleId.ToString() != "3fa85f64-5717-4562-b3fc-2c963f66afa6")
+            {
+                throw new UnauthorizedAccessException();
+            }
 
-            // TODO: добавить все картинки обьявления, а если потом обьявление не добавится, то удалить их
 
-            // TODO: удалить все старые картинки
+            foreach (var imageId in updateRequest.NewAdvertImagesId)
+            {
+                var imageAddRequest = new AdvertImageAddRequest { AdvertId = id, FileId = imageId };
+                await _advertImageRepository.AddAsync(imageAddRequest, cancellation);
+            }
+
+            foreach (var imageId in updateRequest.RemovedAdvertImagesId)
+            {
+                await _advertImageRepository.DeleteByFileIdAsync(imageId, cancellation);
+            }
 
             var updatedDto = await _advertRepository.UpdateAsync(id, updateRequest, cancellation);
+
+
+
+            var user = await _userService.GetById(updatedDto.UserId, cancellation);
+            updatedDto.User = user;
 
             return updatedDto;
         }
@@ -94,6 +162,13 @@ namespace Board.Application.AppData.Contexts.Adverts.Services
 
         public async Task DeleteAsync(Guid id, CancellationToken cancellation)
         {
+            var currentAdvert = await _advertRepository.GetByIdAsync(id, cancellation);
+            var currentUser = await _userService.GetCurrent(cancellation);
+            if (currentUser.Id != currentAdvert.UserId && currentUser.RoleId.ToString() != "3fa85f64-5717-4562-b3fc-2c963f66afa6")
+            {
+                throw new UnauthorizedAccessException();
+            }
+
             await _advertRepository.DeleteAsync(id, cancellation);
         }
     }
